requirements (1).txt
fastapi
uvicorn
anthropic
requests
python-dotenv
python-multipart
jinja2
